+++ apps/laboratory/src/utils/DataUtil.ts
@@ -35,9 +35,14 @@ export const wagmiSdkOptions = [
     description: 'Configuration using wagmi and implementing email and social login'
   },
   {
-    title: 'Permissions',
-    link: '/library/wagmi-permissions/',
-    description: 'Configuration using wagmi and implementing ERC-7715'
+    title: 'Permissions(Sync)',
+    link: '/library/wagmi-permissions-sync/',
+    description: 'Configuration using wagmi and implementing ERC-7715 with passkey'
+  },
+  {
+    title: 'Permissions(Async)',
+    link: '/library/wagmi-permissions-async/',
+    description: 'Configuration using wagmi and implementing ERC-7715 with ecdsa key'
   }
 ]