+++ apps/laboratory/tests/shared/pages/ModalPage.ts
@@ -97,15 +97,15 @@ export class ModalPage {
     await email.deleteAllMessages(emailAddress)
     await this.loginWithEmail(emailAddress)
 
-    let messageId = await email.getLatestMessageId(emailAddress)
-
-    if (!messageId) {
+    const firstMessageId = await email.getLatestMessageId(emailAddress)
+    if (!firstMessageId) {
       throw new Error('No messageId found')
     }
-    let emailBody = await email.getEmailBody(emailAddress, messageId)
+
+    const firstEmailBody = await email.getEmailBody(emailAddress, firstMessageId)
     let otp = ''
-    if (email.isApproveEmail(emailBody)) {
-      const url = email.getApproveUrlFromBody(emailBody)
+    if (email.isApproveEmail(firstEmailBody)) {
+      const url = email.getApproveUrlFromBody(firstEmailBody)
 
       await email.deleteAllMessages(emailAddress)
 
@@ -114,16 +114,18 @@ export class ModalPage {
       await drp.approveDevice()
       await drp.close()
 
-      messageId = await email.getLatestMessageId(emailAddress)
-
-      emailBody = await email.getEmailBody(emailAddress, messageId)
-      if (!email.isApproveEmail(emailBody)) {
-        otp = email.getOtpCodeFromBody(emailBody)
+      const secondMessageId = await email.getLatestMessageId(emailAddress)
+      if (!secondMessageId) {
+        throw new Error('No messageId found')
       }
-    }
 
-    if (otp === '') {
-      otp = email.getOtpCodeFromBody(emailBody)
+      const secondEmailBody = await email.getEmailBody(emailAddress, secondMessageId)
+      if (email.isApproveEmail(secondEmailBody)) {
+        throw new Error('Unexpected approve email after already approved')
+      }
+      otp = email.getOtpCodeFromBody(secondEmailBody)
+    } else {
+      otp = email.getOtpCodeFromBody(firstEmailBody)
     }
 
     await this.enterOTP(otp)
@@ -220,20 +222,17 @@ export class ModalPage {
     await expect(disconnectBtn, 'Disconnect button should be visible').toBeVisible()
     await expect(disconnectBtn, 'Disconnect button should be enabled').toBeEnabled()
     await disconnectBtn.click()
-    await this.page.waitForTimeout(1000)
   }
 
   async sign() {
     const signButton = this.page.getByTestId('sign-message-button')
     await signButton.scrollIntoViewIfNeeded()
-    await this.page.waitForTimeout(500)
     await signButton.click()
-    await this.page.waitForTimeout(1000)
   }
 
-  async signatureRequestFrameShouldVisible() {
+  async signatureRequestFrameShouldVisible(headerText: string) {
     await expect(
-      this.page.frameLocator('#w3m-iframe').getByText('requests a signature'),
+      this.page.frameLocator('#w3m-iframe').getByText(headerText),
       'Web3Modal iframe should be visible'
     ).toBeVisible({
       timeout: 10000
@@ -249,13 +248,18 @@ export class ModalPage {
   }
 
   async approveSign() {
-    await this.signatureRequestFrameShouldVisible()
+    await this.signatureRequestFrameShouldVisible('requests a signature')
     await this.clickSignatureRequestButton('Sign')
   }
 
   async rejectSign() {
-    await this.signatureRequestFrameShouldVisible()
+    await this.signatureRequestFrameShouldVisible('requests a signature')
     await this.clickSignatureRequestButton('Cancel')
+  }
+
+  async approveMultipleTransactions() {
+    await this.signatureRequestFrameShouldVisible('requests multiple transactions')
+    await this.clickSignatureRequestButton('Approve')
   }
 
   async clickWalletUpgradeCard(context: BrowserContext) {
@@ -308,9 +312,9 @@ export class ModalPage {
     await this.page.waitForTimeout(300)
   }
 
-  async updateEmail(mailsacApiKey: string, index: number) {
+  async updateEmail(mailsacApiKey: string) {
     const email = new Email(mailsacApiKey)
-    const newEmailAddress = email.getEmailAddressToUse(index)
+    const newEmailAddress = await email.getEmailAddressToUse()
 
     await this.page.getByTestId('account-button').click()
     await this.page.getByTestId('w3m-account-email-update').click()
@@ -367,6 +371,11 @@ export class ModalPage {
     await this.page.getByTestId('account-button').click()
   }
 
+  async openNetworks() {
+    await this.page.getByTestId('w3m-account-select-network').click()
+    await expect(this.page.getByText('Choose Network')).toBeVisible()
+  }
+
   async openProfileView() {
     await this.page.getByTestId('wui-profile-button').click()
   }
@@ -376,5 +385,46 @@ export class ModalPage {
     await expect(walletFeatureButton).toBeVisible()
 
     return walletFeatureButton
+  }
+
+  async sendCalls() {
+    const sendCallsButton = this.page.getByTestId('send-calls-button')
+    await sendCallsButton.isVisible()
+    await sendCallsButton.click()
+  }
+  async getCallsStatus(batchCallId: string) {
+    const sendCallsInput = this.page.getByTestId('get-calls-id-input')
+    const sendCallsButton = this.page.getByTestId('get-calls-status-button')
+    await sendCallsButton.scrollIntoViewIfNeeded()
+
+    await sendCallsInput.fill(batchCallId)
+    await sendCallsButton.click()
+  }
+
+  async switchAccount() {
+    const switchAccountButton1 = this.page.getByTestId('w3m-switch-address-button-1')
+    await expect(switchAccountButton1).toBeVisible()
+    await switchAccountButton1.click()
+  }
+
+  async getAddress(): Promise<`0x${string}`> {
+    const address = await this.page.getByTestId('w3m-address').textContent()
+    expect(address, 'Address should be present').toBeTruthy()
+
+    return address as `0x${string}`
+  }
+
+  async getChainId(): Promise<number> {
+    const chainId = await this.page.getByTestId('w3m-chain-id').textContent()
+    expect(chainId, 'Chain ID should be present').toBeTruthy()
+
+    return Number(chainId)
+  }
+
+  async getSignature(): Promise<`0x${string}`> {
+    const signature = await this.page.getByTestId('w3m-signature').textContent()
+    expect(signature, 'Signature should be present').toBeTruthy()
+
+    return signature as `0x${string}`
   }
 }