+++ packages/solana/exports/react.tsx
@@ -5,7 +5,7 @@ import { ConstantsUtil } from '@web3modal/scaffold-utils'
 import { getWeb3Modal } from '@web3modal/scaffold-react'
 import { AppKit, WcStoreUtil } from '@web3modal/base'
 import type { AppKitOptions } from '@web3modal/base'
-import { SolanaWeb3JsClient, SolStoreUtil } from '@web3modal/base/adapters/solana/web3js'
+import { SolanaWeb3JsClient } from '@web3modal/base/adapters/solana/web3js'
 import type {
   Chain,
   Provider,